Disable roman numeral parsing until it is fixed
Currently, the roman numeral parsing detects "I ..." messages as the
numeral "I" and doesn't consider an example like "Very", as not being a
valid roman numeral and parses it as "V". I don't have a good fix for
this that doesn't involve requiring a prefix just to send a regular
message so for now it is disabled.